fixed wrong symbol

+++ src/lgnc_directvideo.c
@@ -3,15 +3,15 @@
 
 int LGNC_DIRECTVIDEO_Close()
 {
-    return _LGNC_DIRECTAUDIO_Close();
+    return (int)_LGNC_DIRECTAUDIO_Close();
 }
 
 int LGNC_DIRECTVIDEO_Open(LGNC_VDEC_DATA_INFO_T *info)
 {
-    return _LGNC_DIRECTVIDEO_Open(info);
+    return (int)_LGNC_DIRECTVIDEO_Open(info);
 }
 
 int LGNC_DIRECTVIDEO_Play(void *data, unsigned int size)
 {
-    return _LGNC_DIRECTVIDEO_Play(data, size);
+    return (int)_LGNC_DIRECTVIDEO_Play(data, size);
 }
